updated verilog files
added a bunch of integration files. None of them actually work. I also
made a clock module so we don't have to write one in all the time. I
also added a "double clock" module to double clock the memory. I'm not
sure if this is the correct way to go about that, but I think it should
work if I we can get the wiring right.